+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-xxxxxxxxxxxxxxxxxxxx
+NEXT
 ====================
-- Apify SDK package now prints warning when it's outdated.
+- Added `options.gotoTimeoutSecs` to `PuppeteerCrawler` to enable easier setting of navigation timeouts.
+- `PuppeteerPool` options that were deprecated from the `PuppeteerCrawler` constructor were finally removed.
+  Please use `maxOpenPagesPerInstance`, `retireInstanceAfterRequestCount`, `instanceKillerIntervalSecs`,
+  `killInstanceAfterSecs` and `proxyUrls` via the `puppeteerPoolOptions` object.
+- On the Apify Platform a warning will now be printed when using an outdated `apify` package version.
 
 0.15.3 / 2019-07-29
 ====================