+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 NEXT
 ====================
 - Bump Puppeteer to 1.19.0 and use `{ pipe: true }` again because upstream bug has been fixed.
+- Added `options.gotoTimeoutSecs` to `PuppeteerCrawler` to enable easier setting of navigation timeouts.
+- `PuppeteerPool` options that were deprecated from the `PuppeteerCrawler` constructor were finally removed.
+  Please use `maxOpenPagesPerInstance`, `retireInstanceAfterRequestCount`, `instanceKillerIntervalSecs`,
+  `killInstanceAfterSecs` and `proxyUrls` via the `puppeteerPoolOptions` object.
+- On the Apify Platform a warning will now be printed when using an outdated `apify` package version.
 
 0.15.3 / 2019-07-29
 ====================